refactor: sprite3.png 추가 및 img 디렉토리 assets/Images로 이동
