+++ src/components/Home/HomeStories.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import styled from "styled-components";
 import Card from "UI/Card/Card";
 import Story from "./Story";
-import sprite from "../../Imgs/sprite.png";
+import sprite from "../../assets/Images/sprite.png";
 
 const DUMMY_STORIES = [
     {
